\name{migui}
\alias{migui}
\title{
    Graphical User Interface of the mi Package
}
\description{
    Graphical User Interface of the mi Package
}
\usage{
  migui()
}



\references{
    Yu-Sung Su, Andrew Gelman, Jennifer Hill, Masanao Yajima. \emph{Forthcoming}.
    \dQuote{Multiple Imputation with Diagnostics (mi) in R: Opening Windows into the Black Box}.
    \emph{Journal of Statistical Software}.
}

\author{
  Daniel Lee <bearlee@alum.mit.edu>,
  Yu-Sung Su \email{ys463@columbia.edu}
}

\examples{
  migui()
}

\seealso{
    \code{\link[mi]{mi}}
}


\keyword{manipu}
